requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==2.0.12
Django==3.2.5
django-nose==1.4.7
django-test-without-migrations==0.6
djangorestframework==3.12.4
exceptiongroup==1.3.1
idna==3.11
iniconfig==2.1.0
nose==1.3.7
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
pytest==8.4.2
pytz==2026.1.post1
requests==2.26.0
sqlparse==0.5.5
tomli==2.4.1
typing_extensions==4.15.0
urllib3==1.26.20
